+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2020-01-15
+
+- Updated the package com.gamelovers.configscontainer to version 0.2.0
+- Improved the CsvParser to include special characters like money symbols and dot
+- Added easy selection of the GoogleSheetImporter.asset file. Just go to Tools > Select GoogleSheetImporter.asset. If the GoogleSheetImporter.asset does not exist, it will create a new one in the Assets folder
+
 ## [0.2.1] - 2020-01-15
 
 - Removed Debug.Log lines